Separate UserControllerTest into User and Owner specific tests.
